[N/A] Block Features Organization
