+++ wp-content/plugins/acf-blocks-toolkit/package.json
@@ -20,7 +20,10 @@
     "classnames": "^2.3.2"
   },
   "devDependencies": {
-    "@wordpress/icons": "^9.35.0",
+    "@wordpress/block-editor": "^6.0.0",
+    "@wordpress/components": "^12.0.9",
+    "@wordpress/hooks": "^3.0.0",
+    "@wordpress/i18n": "^4.1.0",
     "@wordpress/scripts": "^26.15.0",
     "prettier": "^3.2.5",
     "webpack-remove-empty-scripts": "^0.8.4"
